Add reporter::prepare_results
This function consolidates &[Mutation] and &[ExecutionResult] into Vec<ExecutedMutant>, looking up code locations in the process.

+++ src/mutation.rs
@@ -5,10 +5,11 @@ use crate::operator::OperatorRegistry;
 use crate::wasmmodule::CallbackType;
 use crate::{config::Config, policy::MutationPolicy, wasmmodule::WasmModule};
 
+#[derive(Debug)]
 pub struct Mutation {
-    pub function_number: u32,
-    pub statement_number: u32,
-    pub offset: u32,
+    pub function_number: u64,
+    pub statement_number: u64,
+    pub offset: u64,
     pub operator: Box<dyn InstructionReplacement>,
 }
